+++ elk/training/train.py
@@ -1,27 +1,26 @@
 """Main training loop."""
 
-import pickle
 import warnings
 from dataclasses import dataclass
 from functools import partial
 from pathlib import Path
-from typing import Literal, Optional, Callable
+from typing import Callable, Literal, Optional
 
 import torch
 from simple_parsing import Serializable, field, subgroups
-from sklearn.metrics import accuracy_score, roc_auc_score
 from torch import Tensor
 
 from elk.extraction.extraction import Extract
 from elk.run import Run
+from elk.training.baseline import (evaluate_baseline, save_baseline,
+                                   train_baseline)
 from elk.utils.typing import assert_type
 
+from ..utils import select_usable_devices
 from .ccs_reporter import CcsReporter, CcsReporterConfig
-from .classifier import Classifier
 from .eigen_reporter import EigenReporter, EigenReporterConfig
 from .reporter import OptimConfig, Reporter, ReporterConfig
 from .train_log import ElicitLog
-from ..utils import select_usable_devices
 
 
 @dataclass
@@ -65,40 +64,6 @@ class Elicit(Serializable):
 class Train(Run):
     cfg: Elicit
 
-    def train_baseline(
-        self,
-        x0: Tensor,
-        x1: Tensor,
-        val_x0: Tensor,
-        val_x1: Tensor,
-        train_labels: Tensor,
-        val_labels: Tensor,
-        device: str,
-    ):
-        # repeat_interleave makes `num_variants` copies of each label, all within a
-        # single dimension of size `num_variants * 2 * n`, such that the labels align
-        # with X.view(-1, X.shape[-1])
-        train_labels_aug = torch.cat(
-            [train_labels, 1 - train_labels]
-        ).repeat_interleave(x0.shape[1])
-        val_labels_aug = (
-            torch.cat([val_labels, 1 - val_labels]).repeat_interleave(x0.shape[1])
-        ).cpu()
-
-        X = torch.cat([x0, x1]).squeeze()
-        d = X.shape[-1]
-        lr_model = Classifier(d, device=device)
-        lr_model.fit_cv(X.view(-1, d), train_labels_aug)
-
-        X_val = torch.cat([val_x0, val_x1]).view(-1, d)
-        with torch.no_grad():
-            lr_preds = lr_model(X_val).sigmoid().cpu()
-
-        lr_acc = accuracy_score(val_labels_aug, lr_preds > 0.5)
-        lr_auroc = roc_auc_score(val_labels_aug, lr_preds)
-
-        return lr_model, lr_auroc, lr_acc
-
     def create_models_dir(self, out_dir: Path):
         lr_dir = None
         lr_dir = out_dir / "lr_models"
@@ -108,10 +73,6 @@ class Train(Run):
         reporter_dir.mkdir(parents=True, exist_ok=True)
 
         return reporter_dir, lr_dir
-
-    def save_baseline(self, lr_dir: Path, layer: int, lr_model: Classifier):
-        with open(lr_dir / f"layer_{layer}.pt", "wb") as file:
-            pickle.dump(lr_model, file)
 
     def train_reporter(
         self,
@@ -152,18 +113,13 @@ class Train(Run):
         )
 
         if not self.cfg.skip_baseline:
-            lr_model, lr_auroc, lr_acc = self.train_baseline(
-                x0,
-                x1,
-                val_x0,
-                val_x1,
-                train_labels,
-                val_labels,
-                device,
-            )
+            lr_model = train_baseline(x0, x1, train_labels, device=device)
+
+            lr_auroc, lr_acc = evaluate_baseline(lr_model, val_x0, val_x1, val_labels)
+
             stats.lr_auroc = lr_auroc
             stats.lr_acc = lr_acc
-            self.save_baseline(lr_dir, layer, lr_model)
+            save_baseline(lr_dir, layer, lr_model)
 
         with open(reporter_dir / f"layer_{layer}.pt", "wb") as file:
             torch.save(reporter, file)
